Suppress future incompatibility lint in CloneToAny impl
There's a new future incompatibility warning that recently started
firing on the code in CloneToAny (see [1]). It's triggering on code that
does a pointer cast from *mut dyn Trait to *mut dyn Trait + Send (or
other auto trait) which is an exact match for what this crate is doing.

The lint doesn't actually apply here though so it is safe to suppress
it. The recommended way to suppress it is to replace the pointer cast
with a transmute and that's what I have done here.

See the new safety comment for a more detailed explanation of why it is
safe.

[1]: https://github.com/rust-lang/rust/issues/127323